--- build.properties
@@ -1,1 +1,0 @@
-discordian_connect_version=8122
