+++ gradle.properties
@@ -1,0 +1,1 @@
+chatChainConnectVersion=8483
