+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'com.minecolonies.discordianbot'
+rootProject.name = 'com.minecolonies.chatchaindc'
