+++ src/main/java/com/minecolonies/chatchaindc/modules/api/config/APIConfigAdapter.java
@@ -1,0 +1,11 @@
+package com.minecolonies.chatchaindc.modules.api.config;
+
+import uk.co.drnaylor.quickstart.config.TypedAbstractConfigAdapter;
+
+public class APIConfigAdapter extends TypedAbstractConfigAdapter.StandardWithSimpleDefault<APIConfig>
+{
+    public APIConfigAdapter()
+    {
+        super(APIConfig.class);
+    }
+}
