+++ src/main/java/com/minecolonies/chatchaindc/modules/api/config/APIConfigReplacements.java
@@ -1,0 +1,14 @@
+package com.minecolonies.chatchaindc.modules.api.config;
+
+/**
+ * Class full of constants that are replaced from the APIConfig.
+ */
+public class APIConfigReplacements
+{
+
+    public static final String CLIENT_NAME = "{client-name}";
+    public static final String CLIENT_TYPE = "{client-type}";
+    public static final String CHANNEL_NAME = "{channel-name}";
+    public static final String USER_NAME = "{user-name}";
+    public static final String USER_MESSAGE = "{user-message}";
+}
