+++ src/main/java/com/minecolonies/chatchaindc/modules/core/config/CoreConfigAdapter.java
@@ -1,0 +1,12 @@
+package com.minecolonies.chatchaindc.modules.core.config;
+
+import uk.co.drnaylor.quickstart.config.TypedAbstractConfigAdapter;
+
+public class CoreConfigAdapter extends TypedAbstractConfigAdapter.StandardWithSimpleDefault<CoreConfig>
+{
+    public CoreConfigAdapter()
+    {
+        // Required thanks to type erasure.
+        super(CoreConfig.class);
+    }
+}
