+++ src/main/java/com/minecolonies/chatchaindc/util/URLParameterStringBuilder.java
@@ -1,0 +1,27 @@
+package com.minecolonies.chatchaindc.util;
+
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.util.Map;
+
+public class URLParameterStringBuilder
+{
+    public static String getParamsString(Map<String, String> params)
+      throws UnsupportedEncodingException
+    {
+        StringBuilder result = new StringBuilder();
+
+        for (Map.Entry<String, String> entry : params.entrySet())
+        {
+            result.append(URLEncoder.encode(entry.getKey(), "UTF-8"));
+            result.append("=");
+            result.append(URLEncoder.encode(entry.getValue(), "UTF-8"));
+            result.append("&");
+        }
+
+        String resultString = result.toString();
+        return resultString.length() > 0
+                 ? resultString.substring(0, resultString.length() - 1)
+                 : resultString;
+    }
+}
