--- src/main/java/com/minecolonies/discordianbot/modules/api/config/APIConfigAdapter.java
@@ -1,11 +1,0 @@
-package com.minecolonies.discordianbot.modules.api.config;
-
-import uk.co.drnaylor.quickstart.config.TypedAbstractConfigAdapter;
-
-public class APIConfigAdapter extends TypedAbstractConfigAdapter.StandardWithSimpleDefault<APIConfig>
-{
-    public APIConfigAdapter()
-    {
-        super(APIConfig.class);
-    }
-}
