--- src/main/java/com/minecolonies/discordianbot/modules/api/config/APIConfigReplacements.java
@@ -1,14 +1,0 @@
-package com.minecolonies.discordianbot.modules.api.config;
-
-/**
- * Class full of constants that are replaced from the APIConfig.
- */
-public class APIConfigReplacements
-{
-
-    public static final String SERVER_NAME = "{server-name}";
-
-    public static final String PLAYER_NAME = "{player-name}";
-
-    public static final String PLAYER_MESSAGE = "{player-message}";
-}
