--- src/main/java/com/minecolonies/discordianbot/util/URLParameterStringBuilder.java
@@ -1,27 +1,0 @@
-package com.minecolonies.discordianbot.util;
-
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
-import java.util.Map;
-
-public class URLParameterStringBuilder
-{
-    public static String getParamsString(Map<String, String> params)
-      throws UnsupportedEncodingException
-    {
-        StringBuilder result = new StringBuilder();
-
-        for (Map.Entry<String, String> entry : params.entrySet())
-        {
-            result.append(URLEncoder.encode(entry.getKey(), "UTF-8"));
-            result.append("=");
-            result.append(URLEncoder.encode(entry.getValue(), "UTF-8"));
-            result.append("&");
-        }
-
-        String resultString = result.toString();
-        return resultString.length() > 0
-                 ? resultString.substring(0, resultString.length() - 1)
-                 : resultString;
-    }
-}
